Feat (CI): Updated CI workflow to work with a db service to work with ecto
Feat (CI): Updated CI workflow to work with a db service to work with ecto

Feat (CI): Updated CI workflow to work with a db service to work with ecto

Feat (CI): Updated CI workflow to work with a db service to work with ecto

Feat (CI): Updated CI workflow to work with a db service to work with ecto

 This is a combination of 7 commits.

Feat (CI): Updated CI workflow to work with NodeJs to install elm and elm-test dependencies, Added a file in frontend app in order to test.

Feat (CI): Updated CI workflow to work with NodeJs to install elm and elm-test dependencies, Added a file in frontend app in order to test.

Feat (CI): Updated CI workflow to work with NodeJs to install elm and elm-test dependencies, Added a file in frontend app in order to test.

Feat (CI): Updated CI workflow to work with NodeJs to install elm and elm-test dependencies, Added a file in frontend app in order to test.

Feat (CI): Updated CI workflow to work with NodeJs to install elm and elm-test dependencies, Added a file in frontend app in order to test.

Feat (CI): Updated CI workflow to work with NodeJs to install elm and elm-test dependencies, Added a file in frontend app in order to test.

Feat (CI): Updated CI workflow to work with NodeJs to install elm and elm-test dependencies, Added a file in frontend app in order to test.